Move sequencer into acquisition clock domain and reduce clockrate
Putting the sequencer in the fx2 clock domain was an awful idea as the
event tagger would latch the various sequencer channels in different
states presumably due to the looser timing constraints in this domain.

Unfortunately a reduction in the acquisition clockrate was necessary to
close the timing of the sequencer logic.